+++ plugins/setup/src/main/java/com/wazuh/setup/SetupPlugin.java
@@ -19,7 +19,7 @@ package com.wazuh.setup;
 import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.service.ClusterService;
-import org.opensearch.common.unit.TimeValue;
+import org.opensearch.common.settings.Setting;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.env.Environment;
@@ -36,13 +36,13 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 
 import com.wazuh.setup.index.Index;
 import com.wazuh.setup.index.IndexStateManagement;
 import com.wazuh.setup.index.StateIndex;
 import com.wazuh.setup.index.StreamIndex;
+import com.wazuh.setup.settings.PluginSettings;
 import com.wazuh.setup.utils.IndexUtils;
 
 /**
@@ -51,7 +51,6 @@ import com.wazuh.setup.utils.IndexUtils;
  */
 public class SetupPlugin extends Plugin implements ClusterPlugin {
 
-    public static final TimeValue TIMEOUT = new TimeValue(30L, TimeUnit.SECONDS);
     private final List<Index> indices = new ArrayList<>();
 
     /** Default constructor */
@@ -115,5 +114,10 @@ public class SetupPlugin extends Plugin implements ClusterPlugin {
         if (localNode.isClusterManagerNode()) {
             this.indices.forEach(Index::initialize);
         }
+    }
+
+    @Override
+    public List<Setting<?>> getSettings() {
+        return List.of(PluginSettings.TIMEOUT, PluginSettings.BACKOFF);
     }
 }